+++ packages/reown_appkit/lib/modal/appkit_modal_impl.dart
@@ -43,6 +43,8 @@ import 'package:reown_appkit/modal/utils/platform_utils.dart';
 import 'package:reown_appkit/modal/constants/key_constants.dart';
 import 'package:reown_appkit/modal/constants/string_constants.dart';
 import 'package:reown_appkit/modal/pages/account_page.dart';
+import 'package:reown_appkit/modal/pages/approve_magic_request_page.dart';
+import 'package:reown_appkit/modal/pages/social_login_page.dart';
 import 'package:reown_appkit/modal/pages/approve_siwe.dart';
 import 'package:reown_appkit/modal/services/analytics_service/analytics_service.dart';
 import 'package:reown_appkit/modal/services/analytics_service/models/analytics_event.dart';
@@ -2063,6 +2065,7 @@ class ReownAppKitModal
     _magicService.onMagicError.subscribe(_onMagicErrorEvent);
     _magicService.onMagicUpdate.subscribe(_onMagicSessionUpdateEvent);
     _magicService.onMagicRpcRequest.subscribe(_onMagicRequest);
+    _magicService.onMagicLoginRequest.subscribe(_onMagicLoginRequest);
     // Coinbase
     _coinbaseService.onCoinbaseConnect.subscribe(_onCoinbaseConnect);
     _coinbaseService.onCoinbaseError.subscribe(_onCoinbaseError);
@@ -2108,6 +2111,7 @@ class ReownAppKitModal
     _magicService.onMagicError.unsubscribeAll();
     _magicService.onMagicUpdate.unsubscribeAll();
     _magicService.onMagicRpcRequest.unsubscribeAll();
+    _magicService.onMagicLoginRequest.unsubscribeAll();
     //
     // Coinbase
     _coinbaseService.onCoinbaseConnect.unsubscribeAll();
@@ -2290,7 +2294,28 @@ extension _EmailConnectorExtension on ReownAppKitModal {
     _notify();
   }
 
+  void _onMagicLoginRequest(MagicSessionEvent? args) {
+    if (args == null) return;
+    final page = SocialLoginPage(socialOption: AppKitSocialOption.Farcaster);
+    if (_isOpen) {
+      _widgetStack.popAllAndPush(page);
+    } else {
+      openModalView(page);
+    }
+  }
+
   void _onMagicRequest(MagicRequestEvent? args) {
+    if (args?.request != null) {
+      if (_widgetStack.containsKey(KeyConstants.approveTransactionPage)) {
+        return;
+      }
+      if (_isOpen) {
+        _widgetStack.push(const ApproveTransactionPage());
+      } else {
+        openModalView(const ApproveTransactionPage());
+      }
+      return;
+    }
     if (args?.result != null) {
       if (args!.result is JsonRpcError && _widgetStack.canPop()) {
         _widgetStack.pop();